Take NA into account when filtering by Distance and Alpha
As it is now, if Distance or Alpha is a NA value, then the entire line
is replaced by NAs instead of being treated adequately. We can then end
up with rows where WatchLenKm is a NA when we just removed them before.
This si probably due to the fact that the expression (NA == "") is equal
to NA and not TRUE/FALSE.