Improvements to distance.filter
Many improvements to the distance.filter to take into account the
updates in the ECSAS.extract function.
- Use %in% instead of == for comparison to take into account NA values.
That is because NA == "" will not return TRUE/FALSE but NA and can cause
unwanted behaviour
- Added a distanceLabel.field argument to indicate the column where the
distance classes are (defaults to Distance for backward compatibility
but now DistanceCode)
- Added a dist2m argument to give the option to perform the class to
numeric conversion of distances. Defaults to TRUE for backward
compatibility but useless now because the conversion is done in
ECSAS.extract
- Added a test to make sure not to perform an uneccsary rbind
considering it is really time consuming on big datasets
- Modified the test for observations in transect to better reflect the
comment associated to the line